feat: apply admin rights for restricting access on some endpoints using admin user decorator

+++ src/station/station.controller.ts
@@ -13,6 +13,7 @@ import { StationService } from './station.service';
 import { CreateStationDto, UpdateStationDto } from './dto/station.dto';
 import { AuthGuard } from '@nestjs/passport';
 import { ApiBearerAuth, ApiTags } from '@nestjs/swagger';
+import { AdminUser } from '@root/auth/admin-user.decorator';
 
 @ApiTags('Station')
 @ApiBearerAuth()
@@ -22,8 +23,13 @@ export class StationController {
   constructor(private readonly stationService: StationService) {}
 
   @Post()
-  create(@Body() createStationDto: CreateStationDto) {
-    return this.stationService.create(createStationDto);
+  create(
+    @Body() createStationDto: CreateStationDto,
+    @AdminUser() isAdmin: boolean,
+  ) {
+    if (isAdmin) {
+      return this.stationService.create(createStationDto);
+    }
   }
 
   @Get()
@@ -36,13 +42,21 @@ export class StationController {
     return this.stationService.findOne(id);
   }
 
-  /* @Patch(':id')
-  update(@Param('id') id: string, @Body() updateStationDto: UpdateStationDto) {
-    return this.stationService.update(+id, updateStationDto);
-  } */
+  @Patch(':id')
+  update(
+    @Param('id') id: string,
+    @Body() updateStationDto: UpdateStationDto,
+    @AdminUser() isAdmin: boolean,
+  ) {
+    if (isAdmin) {
+      return this.stationService.update(id, updateStationDto);
+    }
+  }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
-    return this.stationService.remove(id);
+  remove(@Param('id') id: string, @AdminUser() isAdmin: boolean) {
+    if (isAdmin) {
+      return this.stationService.remove(id);
+    }
   }
 }
